Feat: Admin API 업로드 + 예외처리

+++ src/controllers/admin.controllers.js
@@ -1,4 +1,5 @@
 "use strict";
+const Joi = require("joi");
 // @ts-check
 /**
  * @typedef {import('express').Request} Request
@@ -27,13 +28,31 @@ class AdminController {
    * @param {NextFunction} next
    */
   async registerProducts(req, res, next) {
-    /** @type {{productName: string; desc: string; price: number; image: string | null;}} */
-    const body = req.body;
-    body.image = req.file?.filename || null;
-    await this.adminService.registerProducts(body);
-    res.status(201).json({
-      message: "상품 등록이 완료되었습니다.",
-    });
+    try {
+      const bodySchema = Joi.object({
+        productName: Joi.string().required(),
+        desc: Joi.string().required(),
+        price: Joi.number().required(),
+        image: Joi.string().required(),
+      });
+      /** @type {{productName: string; desc: string; price: number; image: string | null;}} */
+      const body = req.body;
+      body.image = req.file?.filename || null;
+      const validate = await bodySchema.validateAsync(body).catch(() => false);
+      if (!validate) {
+        return res.status(400).json({
+          message: "잘못된 요청입니다.",
+        });
+      }
+      await this.adminService.registerProducts(body);
+      return res.status(201).json({
+        message: "상품 등록이 완료되었습니다.",
+      });
+    } catch (error) {
+      return res.status(500).json({
+        message: "알 수 없는 오류가 일어났습니다.",
+      });
+    }
   }
 
   /**
@@ -43,13 +62,40 @@ class AdminController {
    * @param {NextFunction} next
    */
   async editProducts(req, res, next) {
-    const { productId } = req.params;
-    /** @type {{productName: string; desc: string; price: number; image: string;}} */
-    const body = req.body;
-    await this.adminService.editProducts(Number(productId), body);
-    res.status(201).json({
-      message: "상품 정보가 변경되었습니다.",
+    const paramsSchema = Joi.object({
+      productId: Joi.string().required(),
     });
+    const bodySchema = Joi.object({
+      productName: Joi.string().required(),
+      desc: Joi.string().required(),
+      price: Joi.number().required(),
+      image: Joi.string().required(),
+    });
+    try {
+      const { productId } = req.params;
+      /** @type {{productName: string; desc: string; price: number; image: string;}} */
+      const body = req.body;
+      body.image = req.file?.filename;
+      const paramValidate = await paramsSchema
+        .validateAsync(req.params)
+        .catch(() => false);
+      const bodyValidate = await bodySchema
+        .validateAsync(body)
+        .catch(() => false);
+      if (!paramValidate || !bodyValidate) {
+        return res.status(400).json({
+          message: "잘못된 요청입니다.",
+        });
+      }
+      await this.adminService.editProducts(Number(productId), body);
+      return res.status(201).json({
+        message: "상품 정보가 변경되었습니다.",
+      });
+    } catch (error) {
+      return res.status(500).json({
+        message: "알 수 없는 오류가 일어났습니다.",
+      });
+    }
   }
 
   /**
@@ -59,11 +105,30 @@ class AdminController {
    * @param {NextFunction} next
    */
   async deleteProducts(req, res, next) {
-    const { productId } = req.params;
-    await this.adminService.deleteProducts(Number(productId));
-    res.status(200).json({
-      message: "상품을 삭제하였습니다.",
-    });
+    try {
+      const paramsSchema = Joi.object({
+        productId: Joi.string().required(),
+      });
+
+      const { productId } = req.params;
+      const paramValidate = await paramsSchema
+        .validateAsync(req.params)
+        .catch(() => false);
+
+      if (!paramValidate) {
+        return res.status(400).json({
+          message: "잘못된 요청입니다.",
+        });
+      }
+      await this.adminService.deleteProducts(Number(productId));
+      return res.status(200).json({
+        message: "상품을 삭제하였습니다.",
+      });
+    } catch (error) {
+      return res.status(500).json({
+        message: "알 수 없는 오류가 일어났습니다.",
+      });
+    }
   }
 
   /**
@@ -73,8 +138,14 @@ class AdminController {
    * @param {NextFunction} next
    */
   async getOrderProducts(req, res, next) {
-    const orders = await this.adminService.getOrderProducts();
-    res.status(200).json(orders);
+    try {
+      const orders = await this.adminService.getOrderProducts();
+      return res.status(200).json(orders);
+    } catch (error) {
+      return res.status(500).json({
+        message: "알 수 없는 오류가 일어났습니다.",
+      });
+    }
   }
 
   /**
@@ -84,13 +155,36 @@ class AdminController {
    * @param {NextFunction} next
    */
   async putProductsStatus(req, res, next) {
-    const { productId } = req.params;
-    /** @type {{status: string}} */
-    const { status } = req.body;
-    await this.adminService.putProductsStatus(Number(productId), status);
-    res.status(201).json({
-      message: "상품 상태를 변경하였습니다.",
-    });
+    try {
+      const paramsSchema = Joi.object({
+        productId: Joi.string().required(),
+      });
+      const bodySchema = Joi.object({
+        status: Joi.number().required(),
+      });
+      const { productId } = req.params;
+      const paramValidate = await paramsSchema
+        .validateAsync(req.params)
+        .catch(() => false);
+      const bodyValidate = await bodySchema
+        .validateAsync(req.body)
+        .catch(() => false);
+      if (!paramValidate || !bodyValidate) {
+        return res.status(400).json({
+          message: "잘못된 요청입니다.",
+        });
+      }
+      /** @type {{status: string}} */
+      const { status } = req.body;
+      await this.adminService.putProductsStatus(Number(productId), status);
+      return res.status(201).json({
+        message: "상품 상태를 변경하였습니다.",
+      });
+    } catch (error) {
+      return res.status(500).json({
+        message: "알 수 없는 오류가 일어났습니다.",
+      });
+    }
   }
 
   /**
@@ -100,13 +194,36 @@ class AdminController {
    * @param {NextFunction} next
    */
   async putUserGrade(req, res, next) {
-    const { userId } = req.params;
-    /** @type {{grade: number}} */
-    const { grade } = req.body;
-    await this.adminService.putUserGrade(Number(userId), grade);
-    res.status(201).json({
-      message: "등급이 변경되었습니다.",
-    });
+    try {
+      const paramsSchema = Joi.object({
+        userId: Joi.string().required(),
+      });
+      const bodySchema = Joi.object({
+        grade: Joi.number().required(),
+      });
+      const { userId } = req.params;
+      const paramValidate = await paramsSchema
+        .validateAsync(req.params)
+        .catch(() => false);
+      const bodyValidate = await bodySchema
+        .validateAsync(req.body)
+        .catch(() => false);
+      if (!paramValidate || !bodyValidate) {
+        return res.status(400).json({
+          message: "잘못된 요청입니다.",
+        });
+      }
+      /** @type {{grade: number}} */
+      const { grade } = req.body;
+      await this.adminService.putUserGrade(Number(userId), grade);
+      return res.status(201).json({
+        message: "등급이 변경되었습니다.",
+      });
+    } catch (error) {
+      return res.status(500).json({
+        message: "알 수 없는 오류가 일어났습니다.",
+      });
+    }
   }
 }
